Caches directly from the sbt_toolpath
Instead of caching the download path, we cache directly the toolpath.
We continue first checking that the sbt installation is not already on
the toolpath where we would expect it to be.

Additionally, this combines the install and download phases since we
don't cache the download sbt dir, but the tooldir anymore anyways.